test: separate README messaging policy from protocol internals
